server/requirements.txt
# ── gRPC ──────────────────────────────────────────────────────────────────────
grpcio==1.63.0
grpcio-tools==1.63.0

# ── Protobuf ──────────────────────────────────────────────────────────────────
protobuf==5.27.0

# ── API HTTP ──────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── InfluxDB ──────────────────────────────────────────────────────────────────
# [async] ajoute aiohttp pour le client asynchrone
influxdb-client[async]==1.43.0

# ── Configuration ─────────────────────────────────────────────────────────────
pydantic-settings==2.2.1

# ── Alertes ────────────────────────────────────────────────────────────────────
# Lecture du fichier config/alerts.yaml
pyyaml==6.0.1
# Client HTTP async pour les webhooks Slack/Discord
# (aiohttp est déjà une dépendance transitive de influxdb-client[async],
#  on l'épingle ici pour rendre la dépendance explicite)
aiohttp==3.9.5
